requirements.txt
httpx
uvicorn
starlette
python-dotenv 